3d-model/requirements.txt
PyOpenGL==3.1.0
pywavefront==1.3.0